requirements.txt
deepgram-sdk==3.2.7
openai==1.55.3
python-dotenv==1.0.1

pyaudio==0.2.14
numpy==1.26.4
sounddevice==0.4.7

aiohttp==3.10.5
websockets==12.0
requests==2.32.3